requirements.txt
jupyterlab==4.0.7
matplotlib==3.8.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
oauthlib==3.2.2
requests-oauthlib==1.3.1
sentinelhub==3.9.4